+++ packages/ng-components-demo/src/app/app-dialog.type.ts
@@ -1,0 +1,4 @@
+
+export type TAppDialogFormData = {
+  text: string;
+};