+++ packages/ng-components-demo/src/main.ts
@@ -1,10 +1,13 @@
 import { enableProdMode } from '@angular/core';
-import { bootstrapApplication } from '@angular/platform-browser';
-import { AppComponent } from './app/app.component';
+import { platformBrowser } from '@angular/platform-browser';
+
+import { AppModule } from './app';
 import { environment } from './environments/environment';
+
 
 if (environment.production) {
   enableProdMode();
 }
 
-bootstrapApplication(AppComponent).catch((err) => console.error(err));
+platformBrowser().bootstrapModule(AppModule)
+  .catch(console.error);