+++ packages/ng-dialog/src/lib/dialog-confirm/dialog-confirm-ref.class.ts
@@ -1,5 +1,4 @@
-import { Observable } from 'rxjs';
-import { filter, first } from 'rxjs/operators';
+import { Observable, Subject } from 'rxjs';
 
 import { DialogConfirmComponent } from './dialog-confirm.component';
 
@@ -9,18 +8,20 @@ import { EDialogType } from '../dialog.enum';
 
 export class DialogConfirmRef extends DialogRef {
 
-  readonly type: EDialogType.CONFIRM = EDialogType.CONFIRM;
+  override readonly type: EDialogType.CONFIRM = EDialogType.CONFIRM;
+
+  private readonly _decisionSubject: Subject<boolean> = this._dialogConfirmComponent.decision;
+
+  protected override _maxEvent = 2;
 
   constructor(private readonly _dialogConfirmComponent: DialogConfirmComponent) {
     super(_dialogConfirmComponent);
   }
 
   onBeforeDecision(fn: (decision: boolean) => boolean = (decision: boolean) => true): Observable<boolean> {
-    return this._dialogConfirmComponent.decision
-      .pipe(
-        filter(fn),
-        first()
-      );
+    this._incrementEventCount();
+
+    return this._decisionSubject.pipe(this._checkAllowanceEvent(fn, () => this._decisionSubject.complete()));
   }
 
 }