+++ packages/ng-dialog/src/lib/dialog-form/dialog-form-ref.class.ts
@@ -1,5 +1,4 @@
-import { Observable } from 'rxjs';
-import { filter, first } from 'rxjs/operators';
+import { Observable, Subject } from 'rxjs';
 
 import { DialogFormComponent } from './dialog-form.component';
 
@@ -9,18 +8,20 @@ import { EDialogType } from '../dialog.enum';
 
 export class DialogFormRef<T> extends DialogRef {
 
-  readonly type: EDialogType.FORM = EDialogType.FORM;
+  override readonly type: EDialogType.FORM = EDialogType.FORM;
+
+  private readonly _dataSubject: Subject<T> = this._dialogFormComponent.data;
+
+  protected override _maxEvent = 2;
 
   constructor(private readonly _dialogFormComponent: DialogFormComponent<T>) {
     super(_dialogFormComponent);
   }
 
   onBeforeData(fn: (data: T) => boolean = (_: T) => true): Observable<T> {
-    return this._dialogFormComponent.data
-      .pipe(
-        filter(fn),
-        first()
-      );
+    this._incrementEventCount();
+
+    return this._dataSubject.pipe(this._checkAllowanceEvent(fn, () => this._dataSubject.complete()));
   }
 
 }